don't use long long

+++ src/sys/types/dev_t.c
@@ -1,6 +1,7 @@
 #include <sys/types.h>
 
-typedef unsigned long long int                                            dev_t;
+typedef                                                        unsigned long int
+	dev_t;
 /*
 POSIX(1)
 */
