+++ src/sys/types/nlink_t.c
@@ -1,6 +1,7 @@
 #include <sys/types.h>
 
-typedef unsigned long long int                                          nlink_t;
+typedef                                                        unsigned long int
+	nlink_t;
 /*
 POSIX(1)
 */
