+++ src/sys/types/uid_t.c
@@ -1,6 +1,7 @@
 #include <sys/types.h>
 
-typedef unsigned long long int                                            uid_t;
+typedef                                                        unsigned long int
+	uid_t;
 /*
 POSIX(1)
 */
